+++ Sources/PopupBridge/PopupBridgeConstants.swift
@@ -1,6 +1,9 @@
 import Foundation
 
-class PopupBridgeConstants {
+public class PopupBridgeConstants {
 
     static let callbackURLScheme: String = "sdk.ios.popup-bridge"
+    static let host: String = "popupbridgev1"
+    public static let notificationName = Notification.Name("popupBridgeReturnURL")
+    
 }